add favorites path
